commit for my love

+++ src/ru/vsu/fedosova/model/Player.java
@@ -2,9 +2,9 @@ package ru.vsu.fedosova.model;
 
 public class Player {
     private final String name;
-    private Integer pot;
+    private Double pot;
 
-    public Player(String name, Integer pot) {
+    public Player(String name, Double pot) {
         this.name = name;
         this.pot = pot;
     }
@@ -13,11 +13,11 @@ public class Player {
         return name;
     }
 
-    public Integer getPot() {
+    public Double getPot() {
         return pot;
     }
 
-    public void setPot(Integer pot) {
+    public void setPot(Double pot) {
         this.pot = pot;
     }
 }
